# api/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is the FAIMS Conductor server API documentation.",
    "version": "1.0.0",
    "title": "FAIMS3 API",
    "termsOfService": "http://faims.edu.au/terms/",
    "contact": {"email": "help@faims.edu.au"},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [{"url": "/api"}],
  "components": {
    "securitySchemes": {
      "Auth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "responses": {
      "UnauthorizedError": {"description": "Access token is missing or invalid"}
    }
  },
  "paths": {
    "/hello/": {
      "get": {
        "summary": "Say hello",
        "description": "Return a hello message as JSON",
        "security": [{"Auth": []}],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {"message": {"type": "string"}}
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"}
        }
      }
    },
    "/info": {
      "get": {
        "summary": "Get information",
        "description": "Provides details of the server such as name and description",
        "produces": ["application/json"],
        "security": {"Auth": []},
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/ListingsObject"}
              }
            }
          }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "summary": "Refresh access token",
        "description": "Get a new JWT using a refresh token",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/definitions/PostRefreshTokenInput"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/PostRefreshTokenResponse"}
              }
            }
          }
        }
      }
    },
    "/auth/exchange": {
      "post": {
        "summary": "Exchange token for access/refresh token pair",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/definitions/PostExchangeTokenInput"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/PostExchangeTokenResponse"}
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ListingsObject": {
      "type": "object",
      "properties": {
        "id": {"type": "string"},
        "name": {"type": "string"},
        "conductor_url": {"type": "string"},
        "description": {"type": "string"},
        "prefix": {"type": "string"}
      }
    },
...
    "PostRefreshTokenInput": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string",
          "description": "The refresh token to exchange for a new access token"
        }
      },
      "required": ["refreshToken"]
    },
    "PostRefreshTokenResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "description": "New JWT access token"
        },
        "refreshToken": {"type": "string", "description": "New refresh token"},
        "expiresIn": {
          "type": "number",
          "description": "Token expiration time in seconds"
        }
      },
      "required": ["accessToken", "refreshToken", "expiresIn"]
    },
    "PostExchangeTokenInput": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "description": "Token to exchange (e.g., OAuth token, invite token)"
        },
        "tokenType": {
          "type": "string",
          "enum": ["oauth", "invite", "long-lived"],
          "description": "Type of token being exchanged"
        }
      },
      "required": ["token"]
    },
    "PostExchangeTokenResponse": {
      "type": "object",
      "properties": {
        "accessToken": {"type": "string", "description": "JWT access token"},
        "refreshToken": {
          "type": "string",
          "description": "Refresh token for obtaining new access tokens"
        },
        "expiresIn": {
          "type": "number",
          "description": "Token expiration time in seconds"
        },
        "user": {
          "type": "object",
          "description": "User information",
          "properties": {
            "user_id": {"type": "string"},
            "name": {"type": "string"},
            "email": {"type": "string"}
          }
        }
      },
      "required": ["accessToken", "refreshToken", "expiresIn"]
    }
  }
}